requirements.txt
google-genai
pytest
requests
fastapi
uvicorn
streamlit
pillow
